requirements.txt
fastapi==0.115.6
pandas==2.2.3
pytz==2024.2
PyYAML==6.0.2
rich==13.9.4
selenium==4.13.0
streamlit==1.40.2
uvicorn==0.34.0
webdriver_manager==4.0.2
